fix(audio): 64-bit ring cursors, COM in start(), audio CI on 2022
- PcmRing: use 64-bit free-running cursors so they do not wrap on 32-bit
  builds (a size_t counter wraps after ~4 GiB, corrupting the buffer
  position at the boundary); they stay lock-free on x86/x64.
- WasapiAudioSink: initialize COM in start() and uninitialize in stop()
  on the calling thread rather than at construction, so the constructor
  does no COM/device work and the interfaces are created and released on
  one thread (documented as the required usage).
- ci: run the audio-provisioning Windows jobs on windows-2022 — the
  kernel-driver install (even pnputil's driver-store import) hangs on the
  windows-2025 image, but works on 2022.

+++ src/audio/include/vox/audio/pcm_ring.hpp
@@ -14,6 +14,7 @@
 
 #include <atomic>
 #include <cstddef>
+#include <cstdint>
 #include <span>
 #include <vector>
 
@@ -71,8 +72,11 @@ private:
   static constexpr std::size_t CacheLineSize = 64;
 
   std::vector<std::byte> buffer_;
-  alignas(CacheLineSize) std::atomic<std::size_t> writeCursor_{0};
-  alignas(CacheLineSize) std::atomic<std::size_t> readCursor_{0};
+  // 64-bit free-running counters so they do not wrap (a 32-bit counter would
+  // wrap after ~4 GiB, corrupting the position at the boundary on x86 builds);
+  // 64-bit atomics are still lock-free on the x86/x64 targets.
+  alignas(CacheLineSize) std::atomic<std::uint64_t> writeCursor_{0};
+  alignas(CacheLineSize) std::atomic<std::uint64_t> readCursor_{0};
 };
 
 #if defined(_MSC_VER)
